requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==4.2.17
django-silk==5.3.2
djangorestframework==3.15.2
gprof2dot==2024.6.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
django-cors-headers>=3.7,<4.0  

